Move gateway to core
