+++ knotx-core/src/main/java/io/knotx/gateway/GatewayKnotOptions.java
