+++ knotx-example/knotx-example-app/pom.xml
@@ -69,11 +69,6 @@
     </dependency>
     <dependency>
       <groupId>io.knotx</groupId>
-      <artifactId>knotx-gateway</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.knotx</groupId>
       <artifactId>knotx-example-action-adapter-http</artifactId>
       <version>${project.version}</version>
     </dependency>
